requirements.txt
asgiref==3.5.0
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
localflavor==1.9
Markdown==3.3.6
Pillow==9.1.0
PyMySQL==1.0.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
